imp: save post

+++ src/main/java/com/mvcblogs/blog/models/PostRepository.java
@@ -2,6 +2,6 @@ package com.mvcblogs.blog.models;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PostRepository extends CrudRepository<Post, Integer> {
+public interface PostRepository extends CrudRepository<Post, Long> {
 
 }
